+++ terkin/sensor/system.py
@@ -55,8 +55,89 @@ class SystemTemperature:
 
 
 class SystemBatteryLevel:
+    """
+    Read the battery level by reading the ADC on a pin connected
+    to a voltage divider, which is pin 16 on the expansion board.
+
+    Implementation
+    ==============
+    Written by Dominik Kapusta <https://github.com/ayoy>. Thanks!
+
+    - https://kapusta.cc/2018/02/02/air-quality-monitor-revisited/
+    - https://github.com/ayoy/upython-aq-monitor/blob/lora/lib/adc.py
+
+    Documentation
+    =============
+    - https://forum.pycom.io/topic/3776/adc-use-to-measure-battery-level-vin-level
+    - https://docs.pycom.io/firmwareapi/pycom/machine/adc
+    - https://docs.pycom.io/tutorials/all/adc
+    """
+
+    # How many times to sample the ADC for making a reading.
+    adc_sample_count = const(1000)
+
+    def __init__(self):
+        """
+        Initialized ADC unit.
+        """
+        self.adc = ADC(id=0)
+
     def read(self):
-        pass
+        """
+        Acquire vbatt reading by sampling ADC.
+        """
+
+        # Power on ADC.
+        self.adc.init()
+
+        log.debug('Reading battery level')
+
+        # Sample ADC a few times.
+        adc_channel = self.adc.channel(attn=ADC.ATTN_2_5DB, pin='P16')
+        adc_samples = [0.0] * self.adc_sample_count
+        adc_mean = 0.0
+        i = 0
+        irq_state = disable_irq()
+        while i < self.adc_sample_count:
+            sample = adc_channel()
+            adc_samples[i] = sample
+            adc_mean += sample
+            i += 1
+        enable_irq(irq_state)
+
+        adc_mean /= self.adc_sample_count
+        adc_variance = 0.0
+        for sample in adc_samples:
+            adc_variance += (sample - adc_mean) ** 2
+        adc_variance /= (self.adc_sample_count - 1)
+
+        raw_voltage = adc_channel.value_to_voltage(4095)
+        mean_voltage = adc_channel.value_to_voltage(int(adc_mean))
+        mean_variance = (adc_variance * 10 ** 6) // (adc_mean ** 2)
+
+        # log.debug("ADC readings. count=%u:\n%s" %(self.adc_sample_count, str(adc_samples)))
+        log.debug("SystemBatteryLevel: Mean of ADC readings (0-4095) = %15.13f" % adc_mean)
+        log.debug("SystemBatteryLevel: Mean of ADC voltage readings (0-%dmV) = %15.13f" % (raw_voltage, mean_voltage))
+        log.debug("SystemBatteryLevel: Variance of ADC readings = %15.13f" % adc_variance)
+        log.debug("SystemBatteryLevel: 10**6*Variance/(Mean**2) of ADC readings = %15.13f" % mean_variance)
+
+        voltage_millivolt = (adc_channel.value_to_voltage(int(adc_mean))) * 147 / 47
+        voltage_volt = voltage_millivolt / 1000.0
+
+        # Shut down ADC channel.
+        adc_channel.deinit()
+
+        log.debug('Finished reading battery level: %d', voltage_volt)
+
+        reading = {'system.voltage': voltage_volt}
+        return reading
+
+    def power_off(self):
+        """
+        Shut down ADC.
+        """
+        log.info('Turning off ADC')
+        self.adc.deinit()
 
 
 class SystemUptime: